Bootloader is working and loading the ChibiESC app.
The bootloader can be compiled and flashed with "make boot_flash".
The ChibiESC firmware can be flashed using st-flash --reset write build/chibiesc.elf 0x08010000. This is not required when Eclipse debugging is used.
To debug through the bootloader with Eclipse, the debug startup configuration needs to be changed: "Initial reset. Type:" should be "halt", the initialization commands should be "monitor flash write_image erase "build/chibiesc.elf" 0x08000000 elf", and the "executable offset" should be set to "08000000".